requirements.txt
streamlit 
langchain 
google-generativeai 
tavily-python
dotenv